Basic error checking
Currently, if a repo is inaccessible for any reason, all of repoman ceases to work (telling the user to check if they're connected to the Internet). Additionally, if a repo returns anything other than valid parsable XML, drozer crashes entirely.

This commit introduces 2 error checks: one for `NetworkExceptions` for repos we couldn't fetch, and one for any exception during XML parsing

Uses with a broken repo in their remotes list should now be alerted that there is *a* problem with a specific repo, and they should be able to carry on using repoman